refactor: share Copilot usage parsing
